requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
image==1.5.33
Pillow==9.0.1
pip==21.2.4
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
setuptools==58.1.0
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
tzdata==2021.5
urllib3==1.26.8
